// upvertice-ui/src/app/services/models/page-response-supplier-offer-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SupplierOfferResponse } from '../models/supplier-offer-response';
export interface PageResponseSupplierOfferResponse {
  content?: Array<SupplierOfferResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}