// upvertice-ui/src/app/services/models/supplier-offer-multipart-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Wrapper for multipart supplier offer creation/update request
 */
export interface SupplierOfferMultipartRequest {

  /**
   * Image for the supplier offer (optional)
   */
  image?: Blob;

  /**
   * Supplier offer data as JSON string
   */
  request?: string;
}